[core] Merge GlyphStore and GlyphAtlas
