+++ src/mbgl/layout/symbol_layout.cpp
@@ -4,9 +4,9 @@
 #include <mbgl/renderer/symbol_bucket.hpp>
 #include <mbgl/style/filter_evaluator.hpp>
 #include <mbgl/sprite/sprite_atlas.hpp>
-#include <mbgl/geometry/glyph_atlas.hpp>
+#include <mbgl/text/glyph_atlas.hpp>
 #include <mbgl/text/get_anchors.hpp>
-#include <mbgl/text/glyph_store.hpp>
+#include <mbgl/text/glyph_atlas.hpp>
 #include <mbgl/text/collision_tile.hpp>
 #include <mbgl/util/utf.hpp>
 #include <mbgl/util/token.hpp>
@@ -124,8 +124,8 @@ bool SymbolLayout::hasSymbolInstances() const {
     return !symbolInstances.empty();
 }
 
-bool SymbolLayout::canPrepare(GlyphStore& glyphStore) {
-    if (!layout.textField.value.empty() && !layout.textFont.value.empty() && !glyphStore.hasGlyphRanges(layout.textFont, ranges)) {
+bool SymbolLayout::canPrepare(GlyphAtlas& glyphAtlas) {
+    if (!layout.textField.value.empty() && !layout.textFont.value.empty() && !glyphAtlas.hasGlyphRanges(layout.textFont, ranges)) {
         return false;
     }
 
@@ -137,8 +137,7 @@ bool SymbolLayout::canPrepare(GlyphStore& glyphStore) {
 }
 
 void SymbolLayout::prepare(uintptr_t tileUID,
-                           GlyphAtlas& glyphAtlas,
-                           GlyphStore& glyphStore) {
+                           GlyphAtlas& glyphAtlas) {
     float horizontalAlign = 0.5;
     float verticalAlign = 0.5;
 
@@ -180,7 +179,7 @@ void SymbolLayout::prepare(uintptr_t tileUID,
         layout.textJustify == TextJustifyType::Left ? 0 :
         0.5;
 
-    auto glyphSet = glyphStore.getGlyphSet(layout.textFont);
+    auto glyphSet = glyphAtlas.getGlyphSet(layout.textFont);
 
     for (const auto& feature : features) {
         if (feature.geometry.empty()) continue;
