+++ src/mbgl/renderer/painter.cpp
@@ -17,7 +17,7 @@
 
 #include <mbgl/sprite/sprite_atlas.hpp>
 #include <mbgl/geometry/line_atlas.hpp>
-#include <mbgl/geometry/glyph_atlas.hpp>
+#include <mbgl/text/glyph_atlas.hpp>
 
 #include <mbgl/shader/shaders.hpp>
