+++ src/mbgl/style/observer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <mbgl/text/glyph_store_observer.hpp>
+#include <mbgl/text/glyph_atlas_observer.hpp>
 #include <mbgl/sprite/sprite_atlas_observer.hpp>
 #include <mbgl/style/source_observer.hpp>
 #include <mbgl/map/update.hpp>
@@ -8,7 +8,7 @@
 namespace mbgl {
 namespace style {
 
-class Observer : public GlyphStoreObserver,
+class Observer : public GlyphAtlasObserver,
                  public SpriteAtlasObserver,
                  public SourceObserver {
 public:
