+++ src/mbgl/style/style.hpp
@@ -5,7 +5,7 @@
 #include <mbgl/style/source_observer.hpp>
 #include <mbgl/style/layer_observer.hpp>
 #include <mbgl/style/update_batch.hpp>
-#include <mbgl/text/glyph_store_observer.hpp>
+#include <mbgl/text/glyph_atlas_observer.hpp>
 #include <mbgl/sprite/sprite_atlas_observer.hpp>
 #include <mbgl/map/mode.hpp>
 #include <mbgl/map/zoom_history.hpp>
@@ -24,7 +24,6 @@
 namespace mbgl {
 
 class FileSource;
-class GlyphStore;
 class GlyphAtlas;
 class SpriteAtlas;
 class LineAtlas;
@@ -36,7 +35,7 @@ class Layer;
 class UpdateParameters;
 class QueryParameters;
 
-class Style : public GlyphStoreObserver,
+class Style : public GlyphAtlasObserver,
               public SpriteAtlasObserver,
               public SourceObserver,
               public LayerObserver,
@@ -103,7 +102,6 @@ public:
     void dumpDebugLogs() const;
 
     FileSource& fileSource;
-    std::unique_ptr<GlyphStore> glyphStore;
     std::unique_ptr<GlyphAtlas> glyphAtlas;
     std::unique_ptr<SpriteAtlas> spriteAtlas;
     std::unique_ptr<LineAtlas> lineAtlas;
