+++ src/mbgl/text/glyph_pbf.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <mbgl/text/glyph.hpp>
-#include <mbgl/text/glyph_store.hpp>
 #include <mbgl/util/font_stack.hpp>
 #include <mbgl/util/noncopyable.hpp>
 
@@ -12,15 +11,17 @@
 
 namespace mbgl {
 
+class GlyphAtlas;
+class GlyphAtlasObserver;
 class AsyncRequest;
 class FileSource;
 
 class GlyphPBF : private util::noncopyable {
 public:
-    GlyphPBF(GlyphStore*,
+    GlyphPBF(GlyphAtlas*,
              const FontStack&,
              const GlyphRange&,
-             GlyphStoreObserver*,
+             GlyphAtlasObserver*,
              FileSource&);
     ~GlyphPBF();
 
@@ -31,7 +32,7 @@ public:
 private:
     std::atomic<bool> parsed;
     std::unique_ptr<AsyncRequest> req;
-    GlyphStoreObserver* observer = nullptr;
+    GlyphAtlasObserver* observer = nullptr;
 };
 
 } // namespace mbgl
